+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## v0.3.0 - UNRELEASED
 
+### Added
+
+* Polls models, schemas
+* Polls CRUD functionality
+
 ### Changed
 
 * Orm and migrations to `sqlalchemy` and `alembic`